drivers: modem: gsm_ppp: wait for operator, instead of wait for valid RSSI
To ensure a valid connection, previously we have waited for a valid RSSI.
Now, we wait for a valid registration to an operator.

The waiting for RSSI  was originally a hack to ensure basic functionality.
However, in roaming situations it can take a considerably longer time to
register with an operator than to get a valid RSSI. Therefore, the operator
binding is a stronger criteria to ensure a valid connection.